+++ core-auth-fe/src/core-auth/components/login-mui/LoginPanel.tsx
@@ -6,12 +6,13 @@ import { TabContext, TabList, TabPanel } from '@mui/lab'
 
 import { LoginForm } from './LoginForm'
 import { SignupForm } from './SignupForm'
-import { SocialSignIn, SocialSignInProps } from './SocialSignIn'
+import { SocialSignIn, SocialSignInProps, SocialSignInMode } from './SocialSignIn'
 
 interface Props {
   ctaDisabled?: boolean
   onLogin?: React.FormEventHandler<HTMLFormElement> | undefined
   onRegister?: React.FormEventHandler<HTMLFormElement> | undefined
+  providers?: string[]
   social?: SocialSignInProps['social']
   socialSignInRender?: SocialSignInProps['renderButton']
   onClickSocialConnect?: Function
@@ -29,6 +30,7 @@ export function LoginPanel({
   onLogin,
   onRegister,
   onPasswordReset,
+  providers,
   social,
   socialSignInRender,
   ctaDisabled = false,
@@ -41,29 +43,33 @@ export function LoginPanel({
     setValue(newValue)
   }
 
+  const emailProviderEnabled = providers?.includes('email')
+
   return (
     <Paper sx={{ margin: { xs: 0 }, minHeight: 425, display: 'flex', flexDirection: 'column' }}>
-      <TabContext value={value}>
-        <Box sx={{ borderBottom: 1, borderColor: 'divider' }}>
-          <TabList centered onChange={handleChange} aria-label="Login or Signup">
-            <Tab label="Login" value="1" disabled={!login} />
-            {registration && <Tab label="Sign Up" value="2" />}
-          </TabList>
-        </Box>
-          <TabPanel value="1">
-            <LoginForm
-              ctaDisabled={ctaDisabled}
-              onSubmit={onLogin}
-              onPasswordReset={onPasswordReset}
-            />
-          </TabPanel>
-        {registration && (
-          <TabPanel value="2">
-            <SignupForm onSubmit={onRegister} />
-          </TabPanel>
-        )}
-      </TabContext>
-      <SocialSignIn renderButton={socialSignInRender} social={social} />
+      {emailProviderEnabled && (
+        <TabContext value={value}>
+          <Box sx={{ borderBottom: 1, borderColor: 'divider' }}>
+            <TabList centered onChange={handleChange} aria-label="Login or Signup">
+              <Tab label="Login" value="1" disabled={!login} />
+              {registration && <Tab label="Sign Up" value="2" />}
+            </TabList>
+          </Box>
+            <TabPanel value="1">
+              <LoginForm
+                ctaDisabled={ctaDisabled}
+                onSubmit={onLogin}
+                onPasswordReset={onPasswordReset}
+              />
+            </TabPanel>
+          {registration && (
+            <TabPanel value="2">
+              <SignupForm onSubmit={onRegister} />
+            </TabPanel>
+          )}
+        </TabContext>
+      )}
+      <SocialSignIn mode={emailProviderEnabled ? SocialSignInMode.Footer : SocialSignInMode.Main} renderButton={socialSignInRender} social={social} />
     </Paper>
   )
 }