+++ core-jobs-fe/CHANGELOG.md
@@ -1,5 +1,19 @@
 # @groovestack/jobs
 
+## 0.2.12
+
+### Patch Changes
+
+- Updated dependencies
+  - @groovestack/base@0.2.13
+
+## 0.2.11
+
+### Patch Changes
+
+- Updated dependencies
+  - @groovestack/base@0.2.12
+
 ## 0.2.10
 
 ### Patch Changes