+++ test-core-fe/CHANGELOG.md
@@ -1,5 +1,130 @@
 # test-core-fe
 
+## 0.2.131
+
+### Patch Changes
+
+- Updated dependencies
+  - @groovestack/auth@0.2.25
+
+## 0.2.130
+
+### Patch Changes
+
+- Updated dependencies
+  - @groovestack/auth@0.2.24
+
+## 0.2.129
+
+### Patch Changes
+
+- Updated dependencies
+  - @groovestack/auth@0.2.23
+
+## 0.2.128
+
+### Patch Changes
+
+- Updated dependencies
+  - @groovestack/auth@0.2.22
+
+## 0.2.127
+
+### Patch Changes
+
+- Updated dependencies
+  - @groovestack/webhooks@0.2.16
+
+## 0.2.126
+
+### Patch Changes
+
+- Updated dependencies
+  - @groovestack/auth@0.2.21
+
+## 0.2.125
+
+### Patch Changes
+
+- Updated dependencies
+  - @groovestack/auth@0.2.20
+
+## 0.2.124
+
+### Patch Changes
+
+- Updated dependencies
+  - @groovestack/auth@0.2.19
+
+## 0.2.123
+
+### Patch Changes
+
+- Updated dependencies
+  - @groovestack/auth@0.2.18
+
+## 0.2.122
+
+### Patch Changes
+
+- Updated dependencies
+  - @groovestack/auth@0.2.17
+
+## 0.2.121
+
+### Patch Changes
+
+- Updated dependencies
+  - @groovestack/auth@0.2.16
+
+## 0.2.120
+
+### Patch Changes
+
+- Updated dependencies
+  - @groovestack/auth@0.2.15
+
+## 0.2.119
+
+### Patch Changes
+
+- Updated dependencies
+  - @groovestack/notifications@0.2.33
+  - @groovestack/comments@0.2.15
+  - @groovestack/versions@0.2.15
+
+## 0.2.118
+
+### Patch Changes
+
+- Updated dependencies
+  - @groovestack/base@0.2.13
+  - @groovestack/auth@0.2.14
+  - @groovestack/config@0.2.13
+  - @groovestack/jobs@0.2.12
+
+## 0.2.117
+
+### Patch Changes
+
+- StatusInput gracefully handles undefined save context
+- Updated dependencies
+  - @groovestack/base@0.2.12
+  - @groovestack/auth@0.2.13
+  - @groovestack/comments@0.2.14
+  - @groovestack/config@0.2.12
+  - @groovestack/jobs@0.2.11
+  - @groovestack/notifications@0.2.32
+  - @groovestack/versions@0.2.14
+  - @groovestack/webhooks@0.2.15
+
+## 0.2.116
+
+### Patch Changes
+
+- Updated dependencies
+  - @groovestack/base@0.2.11
+
 ## 0.2.115
 
 ### Patch Changes